Testing of the Dropdown Injection tool
Since the scripting depends on an HTML being there to attach to I have taken two HTML examples, one with and one without dropdowns. Upon opening the html the according script will run from the extensionTest.js file which will log if it was successful in the console of the page.